**RevisionProgress.py**
We are expected to prepare thoroughly for our end-of-year Computer Science exam. To track our revision progress, we can create a simple computer program that calculates the percentage of revision completed. This program will store a list of all the topics we need to revise, ask if we have completed each one with a Yes/No question, and then use our answers to calculate the percentage of topics completed. This helps us get a clear view of our progress and manage our study time effectively.